BACKEND/fably_web/requirements.txt
anyio==4.9.0
bcrypt==3.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
colorama==0.4.6
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2025.3.0
future==1.0.0
gradio_client==1.8.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pycparser==2.22
pymongo==4.7.0
PySocks==1.7.1
python-dotenv==0.19.0
PyYAML==6.0.2
requests==2.32.3
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stripe==11.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
zope.interface==7.2
